feat(torneo): implement complete knockout progression and winner handling
This major update completes the tournament's logical engine:
- Added 'ganadorNombre' to track winners in knockout stages (including penalty shootouts).
- Implemented a generic phase-closing mechanism in the backend to advance winners.
- Updated the Admin Panel with dynamic 'Close Phase' buttons and winner selection in results entry.
- Enhanced Bracket visualization to accurately show advanced teams.

+++ backend/src/models/Partido.ts
@@ -64,11 +64,11 @@ Partido.init(
       allowNull: false,
     },
     golesLocal: {
-      type: DataTypes.TINYINT,
+      type: DataTypes.INTEGER,
       allowNull: true,
     },
     golesVisitante: {
-      type: DataTypes.TINYINT,
+      type: DataTypes.INTEGER,
       allowNull: true,
     },
     ganadorNombre: {
